add --max-fps option
Fixes #54

this limits framerate to a given number, useful for VRR displays.
move the FPS counter code around too as the encoder is now responsible for dropping frames due to fps limiting

this prevents the transform/encode, but not the capture itself, as it's
impossible to know a frames timestamp when we queue it for capture.
luckily, this should be one of the less expensive steps.